Heavily refactor the Addr class.
This removes the ability to implicitly convert from an Addr to an
integer type (you must explicitly convert now), and removes the
ambiguous overloaded operators, as well as some operators that I think
we can live without until there's a pressing reason to add them back.

+++ src/Module.cpp
@@ -85,7 +85,7 @@ const SymbolicExpressionSet& Module::getSymbolicExpressions() const {
 void Module::toProtobuf(MessageType* Message) const {
   nodeUUIDToBytes(this, *Message->mutable_uuid());
   Message->set_binary_path(this->BinaryPath);
-  Message->set_preferred_addr(this->PreferredAddr);
+  Message->set_preferred_addr(static_cast<uint64_t>(this->PreferredAddr));
   Message->set_rebase_delta(this->RebaseDelta);
   Message->set_file_format(static_cast<proto::FileFormat>(this->FileFormat));
   Message->set_isa_id(static_cast<proto::ISAID>(this->IsaID));
